LayoutTests/imported/w3c/web-platform-tests/content-security-policy/inside-worker/dedicated-inheritance-expected.txt

CONSOLE MESSAGE: Refused to connect to http://127.0.0.1:8801/common/text-plain.txt?cross-origin-fetch because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Blocked by Content Security Policy.
CONSOLE MESSAGE: Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.

Harness Error (TIMEOUT), message = null

PASS Filesystem and blob.
PASS Same-origin 'fetch()' in http:?pipe=sub|header(Content-Security-Policy,default-src%20*)
PASS Same-origin 'fetch()' in http:?pipe=sub|header(Content-Security-Policy,connect-src%20*)
PASS Same-origin 'fetch()' in http:
PASS Same-origin 'fetch()' in blob:
FAIL Same-origin 'fetch()' in http:?pipe=sub|header(Content-Security-Policy,connect-src%20%27none%27) promise_test: Unhandled rejection with value: object "TypeError: Not allowed by ContentSecurityPolicy"
PASS Same-origin XHR in http:?pipe=sub|header(Content-Security-Policy,default-src%20*)
PASS Same-origin XHR in http:?pipe=sub|header(Content-Security-Policy,connect-src%20*)
PASS Cross-origin 'fetch()' in http:?pipe=sub|header(Content-Security-Policy,default-src%20*)
PASS Cross-origin XHR in http:?pipe=sub|header(Content-Security-Policy,default-src%20*)
FAIL Same-origin XHR in http:?pipe=sub|header(Content-Security-Policy,connect-src%20%27none%27) promise_test: Unhandled rejection with value: "xhr.open should success."
PASS Same-origin => cross-origin 'fetch()' in http:?pipe=sub|header(Content-Security-Policy,default-src%20*)
PASS Cross-origin 'fetch()' in http:?pipe=sub|header(Content-Security-Policy,connect-src%20*)
PASS Same-origin XHR in http:
PASS Cross-origin XHR in http:?pipe=sub|header(Content-Security-Policy,connect-src%20*)
PASS Cross-origin 'fetch()' in http:?pipe=sub|header(Content-Security-Policy,connect-src%20%27none%27)
PASS Same-origin => cross-origin 'fetch()' in http:?pipe=sub|header(Content-Security-Policy,connect-src%20*)
PASS Same-origin XHR in blob:
PASS Cross-origin XHR in http:?pipe=sub|header(Content-Security-Policy,connect-src%20%27none%27)
PASS Same-origin => cross-origin 'fetch()' in http:?pipe=sub|header(Content-Security-Policy,connect-src%20%27none%27)
PASS Cross-origin 'fetch()' in http:
PASS Cross-origin XHR in http:
PASS Cross-origin 'fetch()' in blob:
PASS Same-origin => cross-origin 'fetch()' in http:
PASS Cross-origin XHR in blob:
PASS Same-origin => cross-origin 'fetch()' in blob:
